Improve session handling (fixes issues related to bug 440331)
When the ClientSession ends, the client notification poller (that was
previously started in execLoadSession) should be stopped. Otherwise, a
new server session might be created after logout (see bug 440331).

In the Swing applications, the NetActivator needs to be installed on
startup. Otherwise, no CookieManager is available and the server keeps
'forgetting' the user's HTTP session and creates a new session for every
request. This is not really necessary when running without server
(footless mode), but it doesn't hurt either.